# server/docs/openapi.yaml
openapi: 3.0.0
info:
  description: REST API for the NDHU campus second-hand textbook marketplace.
  title: NDHU Second-Hand Book Store API
  termsOfService: http://swagger.io/terms/
  contact:
    name: SHBS Dev Team
  license:
    name: MIT
  version: "1.0"
paths:
  /auth/login:
    post:
      description: Authenticates a user with email and password, returns a JWT token
      tags:
        - Auth
      summary: Login
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/model.SwaggerLoginRequest"
        description: Login credentials
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerLoginResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
...
  /images:
    post:
      security:
        - BearerAuth: []
      description: Registers an image metadata record in the database after upload to S3
      tags:
        - Images
      summary: Register image
      operationId: registerImage
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/model.SwaggerRegisterImageRequest"
        description: Image metadata
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.ImageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "422":
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
  "/images/{id}":
    get:
      description: Returns image metadata by ID
      tags:
        - Images
      summary: Get image
      operationId: getImage
      parameters:
        - description: Image ID
          name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.ImageResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/model.SwaggerErrorResponse"
...
servers:
  - url: /api/v1
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    model.ImageResponse:
      type: object
      properties:
        cdn_url:
          type: string
        id:
          type: string
...
    model.SwaggerErrorResponse:
      type: object
      properties:
        error:
          type: string
    model.SwaggerLoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    model.SwaggerLoginResponse:
      type: object
      properties:
        expires_at:
          type: string
        token:
          type: string
        user:
          $ref: "#/components/schemas/model.User"
...
    model.SwaggerRegisterImageRequest:
      type: object
      properties:
        cdn_url:
          type: string
        s3_key:
          type: string
...
    model.User:
      type: object
      properties:
        avatar_image_id:
          type: string
        cas_id:
          type: string
        created_at:
          type: string
        email:
          type: string
        email_verified:
          type: boolean
        id:
          type: string
        is_admin:
          type: boolean
        name:
          type: string
        updated_at:
          type: string